searchResult state form SearchBook lifted to App
The idea is to focus searchbook to... search books, once the books
have been found they will be passed to App to make the merging with the
current books more streamlined.

If the result is not that positive Searchbook compoenent can display an
error or no book missing without App needing to know much about it.